Add support for double number format (#526)
Make `f64` correctly to be tokenized as `double` within generated 
OpenAPI spec.

---------

+++ utoipa-gen/tests/schema_derive_test.rs
@@ -382,7 +382,7 @@ fn derive_enum_with_comments_success() {
 #[test]
 fn derive_struct_unnamed_field_single_value_type_success() {
     let point = api_doc! {
-        struct Point(f64);
+        struct Point(f32);
     };
 
     assert_value! {point=>
@@ -403,7 +403,7 @@ fn derive_struct_unnamed_fields_tuple_with_same_type_success() {
     assert_value! {point=>
         "type" = r#""array""#, "Point type"
         "items.type" = r#""number""#, "Point items type"
-        "items.format" = r#""float""#, "Point items format"
+        "items.format" = r#""double""#, "Point items format"
         "items.description" = r#""Contains x and y coordinates\n\nCoordinates are used to pinpoint location on a map""#, "Point items description"
         "maxItems" = r#"2"#, "Wrapper max items"
         "minItems" = r#"2"#, "Wrapper min items"
@@ -2805,7 +2805,7 @@ fn derive_struct_with_rust_decimal_with_type_override() {
         "properties.id.type" = r#""integer""#, "Post id type"
         "properties.id.format" = r#""int32""#, "Post id format"
         "properties.rating.type" = r#""number""#, "Post rating type"
-        "properties.rating.format" = r#""float""#, "Post rating format"
+        "properties.rating.format" = r#""double""#, "Post rating format"
     }
 }
